moved main views into views folder
