+++ src/components/public_views/MemberServices.js
@@ -1,25 +1,25 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 import Helmet from 'react-helmet';
-import SponsorImageOnly from './sponsor/SponsorImageOnly';
-import NoItems from './application/NoItems';
+import MemberServiceSingle from '../member_service/MemberServiceSingle';
+import NoItems from '../application/NoItems';
 
-class Sponsors extends Component {
+class MemberServices extends Component {
     constructor(props) {
         super(props);
         this.state = { 
-            sponsors: []
-        }
+            member_services: []
+         }
     }
 
     componentDidMount() {
-        const url = 'http://localhost:3001/api/v1/sponsors/active';
+        const url = 'http://localhost:3001/api/v1/member_services/active';
     
         axios.get(url)
         .then((response) => {
           console.log(response)
           this.setState({
-            sponsors: response.data
+            member_services: response.data
           })
         })
         .catch((error) => {
@@ -33,32 +33,32 @@ class Sponsors extends Component {
         return ( 
             <div>
                 <Helmet>
-                    <title>BOTSC | Sponsors</title>
+                    <title>BOTSC | Member Services</title>
                 </Helmet>
                 <div className="page-heading">
                     <div className="container">
                         <div className="row">
                             <div className="col-md-10 offset-md-1">
-                                <h1 className="page-heading__title">Sponsors</h1>
+                                <h1 className="page-heading__title">Member <span className="highlight"> Services</span></h1>
                             </div>
                         </div>
                     </div>
                 </div>
                 <div className="site-content">
                     <div className="container">
-                        {this.state.sponsors.length ?
+                        {this.state.member_services.length ?
                             <div className="sponsors-grid row">
-                                {this.state.sponsors.map( sponsor => (
-                                    <div key={sponsor.id} className="content col-lg-4 col-sm-6">
-                                        <SponsorImageOnly
-                                            sponsor={sponsor}
+                                {this.state.member_services.map( member_service => (
+                                    <div key={member_service.id} className="content col-lg-4 col-sm-6">
+                                        <MemberServiceSingle
+                                            member_service={member_service}
                                         />
                                     </div>
                                 ))}
                             </div>
-                        :
-                            <NoItems
-                                item="Sponsors"
+                        : 
+                            <NoItems 
+                                item="Member Services"
                             />
                         }
                     </div>
@@ -68,4 +68,4 @@ class Sponsors extends Component {
     }
 }
  
-export default Sponsors;
+export default MemberServices;
