+++ src/components/public_views/Team.js
@@ -1,25 +1,25 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 import Helmet from 'react-helmet';
-import NoItems from './application/NoItems';
-import TeamCard from './teams/TeamCard';
+import NoItems from '../application/NoItems';
+import PlayerSmall from '../widgets/PlayerSmall';
 
-class Division extends Component {
+class Team extends Component {
     constructor(props) {
         super(props);
         this.state = { 
-            division: {}
+            team: {}
         }
     }
 
     componentDidMount() {
-        const url = 'http://localhost:3001/api/v1/divisions/' + this.props.division_id + '/with_teams_full';
+        const url = 'http://localhost:3001/api/v1/teams/' + this.props.team_id + '/full';
     
         axios.get(url)
         .then((response) => {
           console.log(response)
           this.setState({
-            division: response.data
+            team: response.data
           })
         })
         .catch((error) => {
@@ -30,13 +30,13 @@ class Division extends Component {
     }
 
     componentWillReceiveProps(newProps) {
-        const url = 'http://localhost:3001/api/v1/divisions/' + newProps.division_id + '/with_teams_full';
+        const url = 'http://localhost:3001/api/v1/teams/' + newProps.team_id + '/full';
     
         axios.get(url)
         .then((response) => {
           console.log(response)
           this.setState({
-            division: response.data
+            team: response.data
           })
         })
         .catch((error) => {
@@ -50,35 +50,47 @@ class Division extends Component {
         return ( 
             <div>
                 <Helmet>
-                    <title>{"BOTSC | " + this.state.division.name + " Division"}</title>
+                    <title>{"BOTSC | " + this.state.team.name}</title>
                 </Helmet>
                 <div className="page-heading">
                     <div className="container">
                         <div className="row">
                             <div className="col-md-10 offset-md-1">
-                                <h1 className="page-heading__title">{this.state.division.name}<span className="highlight"> Division</span></h1>
+                                <h1 className="page-heading__title">{this.state.team.name}</h1>
                             </div>
                         </div>
                     </div>
                 </div>
                 <div className="site-content">
                     <div className="container">
-                        {this.state.division.teams ?
+                        {this.state.team.members ?
                             <div className="row">
-                                <h3 className="col-md-12">Teams</h3>
                                 <div className="col-md-12 team-roster team-roster--card-compact js-team-roster--card-compact row">
-                                    {this.state.division.teams.map( team => (
-                                        <div key={team.id} className="content col-lg-4 col-sm-6">
-                                            <TeamCard
-                                                team={team}
+                                    {this.state.team.convenor ?
+                                        <div className="content col-lg-4">
+                                            <PlayerSmall 
+                                                player={this.state.team.convenor}
+                                                convenor={true}
                                             />
                                         </div>
+                                    : null }
+                                    {this.state.team.members.map( player => (
+                                        <>
+                                            {player.id !== this.state.team.convenor.id ?
+                                                <div className="content col-lg-4">
+                                                    <PlayerSmall 
+                                                        player={player}
+                                                        convenor={false}
+                                                    />
+                                                </div>
+                                            : null }
+                                        </>
                                     ))}
                                 </div>
                             </div>
                         :
                             <NoItems
-                                item="Teams"
+                                item="Players"
                             />
                         }
                     </div>
@@ -88,4 +100,4 @@ class Division extends Component {
     }
 }
  
-export default Division;
+export default Team;
