+++ src/components/widgets/Leaderboard.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { Link } from '@reach/router';
+import { Link } from 'react-router-dom';
 
 class Leaderboard extends Component {
     constructor(props) {
